// SimERP/ClientApp/src/app/lists/aggregatecosts/model/aggregatecosts.ts
export class AggregateCosts {
    Id: number;
    LoaiCp: string = "";
    ChiPhi?: number;
    DienGiai: string = "";
    ApplyDate?: Date;
    Notes: string;
    Description: string;
    CreatedDate?: Date;
    CreatedBy?: number;
    ModifyDate?: Date;
    ModifyBy?: number;
    IsActive: boolean = true;
    IsDelete: boolean = false;
    SearchString: string;
    SortOrder?: number;
}